weight tying to post load hook
