+++ tensorrt_llm/_torch/auto_deploy/config/default.yaml
@@ -100,9 +100,6 @@ transforms:
   ############################################################################################
   # RUN POST-LOAD FUSION AND OPTIMIZATIONS
   ############################################################################################
-  sync_tied_weights:
-    stage: post_load_fusion
-    run_per_gm: false
   fuse_gemms:
     stage: post_load_fusion
     enabled: false # TODO: https://github.com/NVIDIA/TensorRT-LLM/issues/4674 this is causing OOMs
